added ability to create custom scrapers
